refactor: extract RocketChatSubscriptionClient from RocketChatService
